Added raw code files due to a GitHub error
We committed our code changes on November 15 at 11:43 PM, but the commit never finished processing, not even by 12:20 AM. Since the hackathon deadline was extended, we tried again today, November 16. We started the commit at 9:58 AM, and as of 6:09 PM, it is still stuck loading. Because of this, I decided to upload the raw code files in hopes that it will finally go through.

If you need proof, feel free to contact @abjectbrake on Discord. I recorded a video yesterday in case the issue continued, which it did, so I can send it over.